Update cache_fill_FSM.v
changes to fsm as per phase 3 brief. from my understanding the fsm needed to be changed although if this was incorrect my last version scored well in HW7 and could be restored. This design using 2 counters one to count the amount of reads performed and one to track how many responses